move link and series config to config folder
