+++ components/x-podcast-launchers/src/config/series-ids.js
@@ -1,0 +1,5 @@
+import { brand } from '@financial-times/n-concept-ids';
+
+export default new Map([
+	[brand.rachmanReviewPodcast, 'therachmanreview']
+]);
