--- components/x-podcast-launchers/src/map-concept-to-acast-series.js
@@ -1,9 +1,0 @@
-import { brand } from '@financial-times/n-concept-ids';
-
-const acastSeriesMap = new Map([
-	[brand.rachmanReviewPodcast, 'therachmanreview']
-]);
-
-export default function mapConceptToAcastSeries(conceptId) {
-	return acastSeriesMap.get(conceptId);
-}
